requirements.txt
arrow==0.10.0
backports.weakref==1.0rc1
binaryornot==0.4.4
bleach==2.0.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
cookiecutter==1.5.1
cycler==0.10.0
decorator==4.1.2
descartes==1.1.0
entrypoints==0.2.3
future==0.16.0
GDAL==2.2.2
geojson==2.3.0
geomet==0.2.0.post2
h5py==2.7.1
html2text==2017.10.4
html5lib==0.999999999
idna==2.6
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.2
jedi==0.11.0
Jinja2==2.9.6
jinja2-time==0.2.0
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
Keras==2.0.8
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.0.2
mistune==0.7.4
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.0
notebook==5.2.0
numpy==1.13.1
olefile==0.44
opencv-python==3.3.0.10
pandas==0.20.3
pandocfilters==1.4.2
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
poyo==0.4.1
prompt-toolkit==1.0.15
protobuf==3.4.0
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.1
requests==2.18.4
scikit-image==0.13.1
scikit-learn==0.19.0
scipy==0.19.1
sen2cor==2.4.0
sentinelsat==0.12.1
Shapely==1.6.1
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorflow-gpu==1.2.1
tensorflow-tensorboard==0.1.6
terminado==0.6
testpath==0.3.1
tifffile==0.12.1
tornado==4.5.2
tqdm==4.19.4
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
whichcraft==0.4.1
widgetsnbextension==3.0.4
